Rename file AudioContent => TTSResponseBody. Add translate json data classes
