+++ Assets/Scripts/Translate.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ca6b1d8e11eab00498b7294cab96721b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
