+++ Assets/Scripts/Translate/JsonData.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bb9fe24fb8a70e3419023d2a8e0a8b04
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
