+++ Assets/Scripts/Translate/JsonData/TranslateRequestBody.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d858d56f282926d43aada506e00144b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
