+++ Assets/Scripts/Translate/JsonData/TranslateResponseBody.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace Scripts.Translate.JsonData
+{
+    [Serializable]
+    class TranslateResponseBody
+    {
+        public List<TranslatedText> translations;
+    }
+}
