+++ Assets/Scripts/Translate/JsonData/TranslateResponseBody.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00c402bf1ce37404f83249b009c3950b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
