+++ Assets/Scripts/Translate/JsonData/TranslatedText.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace Scripts.Translate.JsonData
+{
+    [Serializable]
+    class TranslatedText
+    {
+        public string translatedText;
+    }
+}
