+++ Assets/Scripts/Translate/JsonData/TranslatedText.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 17f919af2257ed640b2e14d9c51fffcb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
