+++ Assets/Scripts/Translate/Translate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 970e9fd0f7565cd41bc460716f24d4da
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
